pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["d.krasilnikov <d.i.krasilnikov@tinkoff.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
pyTelegramBotAPI = "^4.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
